Clean up 'Send To Contract Transparent' gadget
This includes grouping all of the gadgets for the inputs
together in one iteration, as well as for the outputs.
Besides this, instead of taking in a rest value, which will be sent
to the contract, it is automatically calculated by use of an
accumulator variable, which tallies up all of the input values,
and subtracts all of the output values. It is then passed to the balance
gadget, at the end of the function.